+++ tests/app/config/config.php
@@ -1,5 +1,5 @@
 <?php
-return OctoLab\Cilex\Config\Util\ArrayHelper::merge(
+return OctoLab\Common\Config\Util\ArrayHelper::merge(
     include 'parameters.php',
     include 'component/config.php',
     [